chore: typescript rewrite
