+++ formidable/common/resources/js/bootstrap.ts
