+++ formidable/common/webpack.config.js
