+++ formidable/presets/react/files/resources/js/app.tsx
@@ -1,0 +1,18 @@
+require('./bootstrap');
+
+import React from 'react'
+import { createRoot } from "react-dom/client"
+import { createInertiaApp } from '@inertiajs/inertia-react'
+import { InertiaProgress } from '@inertiajs/progress'
+
+const appName = window.document.getElementsByTagName('title')[0]?.innerText || 'Formidable';
+
+createInertiaApp({
+    title: (title) => `${title} - ${appName}`,
+    resolve: (name) => import(`./Pages/${name}`),
+    setup({ el, App, props }) {
+        createRoot(el).render(<App {...props} />)
+    },
+});
+
+InertiaProgress.init({ color: '#4B5563' })
