+++ formidable/presets/react/files/webpack.mix.js
@@ -12,12 +12,13 @@ const mix = require('laravel-mix');
  */
 
 mix.setPublicPath('./public')
-    .js('resources/js/app.js', './public/js')
+    .ts('resources/js/app.tsx', './public/js')
     .react()
     .postCss('resources/css/app.css', './public/css', [
         //
     ])
-    .webpackConfig(require('./webpack.config'));
+    .webpackConfig(require('./webpack.config'))
+    .disableNotifications();
 
 if (mix.inProduction()) {
     mix.version();
