+++ formidable/presets/react/preset.json
@@ -12,6 +12,9 @@
             "@inertiajs/inertia": "^0.11.1",
             "@inertiajs/inertia-react": "^0.8.1",
             "@inertiajs/progress": "^0.2.7",
+            "@types/node": "^18.11.17",
+            "@types/react": "^18.0.26",
+            "@types/react-dom": "^18.0.10",
             "axios": "^1.2.2",
             "laravel-mix": "^6.0.49",
             "lodash": "^4.17.21",
