+++ formidable/presets/vue/preset.json
@@ -11,6 +11,7 @@
             "@inertiajs/inertia": "^0.11.1",
             "@inertiajs/inertia-vue3": "^0.6.0",
             "@inertiajs/progress": "^0.2.7",
+            "@types/node": "^18.11.17",
             "@vue/compiler-sfc": "^3.2.45",
             "axios": "^1.2.2",
             "laravel-mix": "^6.0.49",
